+++ azplugins/pair.py
@@ -264,6 +264,90 @@ class lj124(hoomd.md.pair.pair):
 
         return _hoomd.make_scalar2(lj1, lj2)
 
+class ashbaugh24(hoomd.md.pair.pair):
+    R""" generalized Ashbaugh-Hatch 48-24 potential
+
+    Args:
+        r_cut (float): Default cutoff radius (in distance units).
+        nlist (:py:mod:`hoomd.md.nlist`): Neighbor list
+        name (str): Name of the force instance.
+
+    :py:class:`ashbaugh24` is a Lennard-Jones 48-24 potential, implemented as described by
+    `L. Rovigatti, B. Capone, C. Likos, Nanoscale, 8 (2016) <http://dx.doi.org/10.1039/C5NR04661K>`_.
+    The potential has a purely repulsive (Weeks-Chandler-Andersen like) core, with a
+    parameter :math:`\lambda` setting the strength of the attractive tail.
+    When :math:`\lambda` is 0, the potential is purely repulsive.
+    When :math:`\lambda` is 1, the potential a generalized Lennard-Jones potential:
+    .. math::
+        :nowrap:
+
+        \begin{eqnarray*}
+        V(r)  = & V_{\mathrm{LJ,48-24}}(r, \varepsilon, \sigma) + (1-\lambda)\varepsilon & r < (2)^{1/24}\sigma \\
+              = & \lambda V_{\mathrm{LJ,48-24}}(r, \varepsilon, \sigma) & (2)^{1/24}\sigma \ge r < r_{\mathrm{cut}} \\
+              = & 0 & r \ge r_{\mathrm{cut}}
+        \end{eqnarray*}
+
+    Here, :math:`V_{\mathrm{LJ,48-24}}(r,\varepsilon,\sigma)` is a Lennard-Jones potential with
+    parameters :math:`\varepsilon`, and :math:`\sigma`:
+    .. math::
+        :nowrap:
+
+        \begin{eqnarray*}
+        V_{\mathrm{LJ,48-24}}  = 4 \varepsilon \left(\left(\frac{\sigma}{r}\right)^{48} - \left(\frac{\sigma}{r}\right)^{24}\right)
+        \end{eqnarray*}
+
+    See :py:class:`hoomd.md.pair.pair`
+    for details on how forces are calculated and the available energy shifting and smoothing modes.
+    Use :py:meth:`pair_coeff.set <coeff.set>` to set potential coefficients.
+
+    The following coefficients must be set per unique pair of particle types:
+
+    - :math:`\varepsilon` - *epsilon* (in energy units)
+    - :math:`\sigma` - *sigma* (in distance units)
+    - :math:`\lambda` - *lam* (unitless) - scale factor for attraction
+    - :math:`r_{\mathrm{cut}}` - *r_cut* (in distance units)
+      - *optional*: defaults to the global r_cut specified in the pair command
+    - :math:`r_{\mathrm{on}}`- *r_on* (in distance units)
+      - *optional*: defaults to the global r_cut specified in the pair command
+
+    Example::
+
+        nl = hoomd.md.nlist.cell()
+        lj = azplugins.pair.ashbaugh24(r_cut=3.0, nlist=nl)
+        lj.pair_coeff.set('A', 'A', epsilon=1.0, sigma=1.0, lam=0.0)
+        lj.pair_coeff.set(['A','B'], 'B', epsilon=1.0, sigma=1.0, lam=0.5, r_cut=3.0, r_on=2.5)
+
+    """
+    def __init__(self, r_cut, nlist, name=None):
+        hoomd.util.print_status_line()
+
+        # initialize the base class
+        hoomd.md.pair.pair.__init__(self, r_cut, nlist, name)
+
+        # create the c++ mirror class
+        if not hoomd.context.exec_conf.isCUDAEnabled():
+            self.cpp_class = _azplugins.PairPotentialAshbaugh24
+        else:
+            self.cpp_class = _azplugins.PairPotentialAshbaugh24GPU
+            self.nlist.cpp_nlist.setStorageMode(_md.NeighborList.storageMode.full)
+        self.cpp_force = self.cpp_class(hoomd.context.current.system_definition, self.nlist.cpp_nlist, self.name)
+
+        hoomd.context.current.system.addCompute(self.cpp_force, self.force_name)
+
+        # setup the coefficent options
+        self.required_coeffs = ['epsilon', 'sigma', 'lam']
+
+    def process_coeff(self, coeff):
+        epsilon = coeff['epsilon']
+        sigma = coeff['sigma']
+        lam = coeff['lam']
+
+        lj1 = 4.0 * epsilon * math.pow(sigma, 48.0)
+        lj2 = 4.0 * epsilon * math.pow(sigma, 24.0)
+        rwcasq = math.pow(2.0, 1.0/12.0) * sigma**2
+        wca_shift = epsilon * (1. - lam)
+        return _azplugins.make_ashbaugh_params(lj1, lj2, lam, rwcasq, wca_shift)
+
 class lj96(hoomd.md.pair.pair):
     R""" LJ 9-6 potential